+++ src/routes/(app)/eventos/[slug]/+page.svelte
@@ -94,12 +94,11 @@
     content={`${$eventData?.event?.seo_description}`}
   />
   <meta property="og:url" content={`${$page.url.href}`} />
-  <script src="https://www.mercadopago.com/v2/security.js" ></script>
 </svelte:head>
 
 {#if $eventData?.event}
   <div class="xl:container xl:mx-auto min-w-[350px]">
-    <div class="container xl:mx-auto min-w-[350px] mx-auto px-4 mt-20 h-min">
+    <div class="container xl:mx-auto min-w-[350px] mx-auto mt-20 h-min">
       <div
         id="$eventData?.event"
         class="h-fit flex flex-col lg:flex-row md:gap-4 lg:gap-0"
@@ -120,7 +119,7 @@
             alt="Afiche del próximo evento"
           />
         </div>
-        <div class="z-10 relative -top-20 lg:top-0 xl:top-20 lg:-left-12">
+        <div class="z-10 relative -top-20 lg:top-0 lg:-left-12 px-4 lg:px-0">
           <div class="mt-4 lg:mt-0">
             <div class="mx-auto text-base lg:ml-auto lg:mr-0">
               <!-- EVENTO PASADO -->
@@ -427,4 +426,3 @@
 {:else}
   <h1>Ups! no encontramos el evento que buscas.</h1>
 {/if}
-<div id="wallet_container"></div>